refactor: 로또 이익률 계산 수행을 LottoCustomer로 이동
